Edited copy script, also readme needs to be a txt

+++ assn1/copy.sh
@@ -1,3 +1,8 @@
-cp client.py client
-scp client rofael@moat.cs.uab.edu:cs334/lab1
-rm client
+echo "Make sure directory ~/cs334/assn1 exists in your moat"
+if [ "$1" == "" ]; then
+    echo "Specify username for moat"
+else
+    cp client.py client
+    scp client "$1"@moat.cs.uab.edu:cs334/assn1
+    rm client
+fi
